Implementing permutation class
