+++ src/symfc/basis_sets/basis_sets_O1.py
@@ -9,7 +9,7 @@ from scipy.sparse import csr_array
 
 from symfc.eig_solvers.api_eig_tools import eigsh_projector
 from symfc.spg_reps import SpgRepsO1
-from symfc.utils.coset_tools_O1 import get_compr_coset_reps_sum
+from symfc.utils.coset_tools_O1 import get_compr_coset_projector_O1
 from symfc.utils.translation_tools_O1 import compressed_projector_sum_rules
 from symfc.utils.utils import SymfcAtoms
 from symfc.utils.utils_O1 import (
@@ -102,7 +102,7 @@ class FCBasisSetO1(FCBasisSetBase):
     def run(self) -> FCBasisSetO1:
         """Compute compressed force constants basis set."""
         c_trans = self._get_c_trans()
-        coset_reps_sum = get_compr_coset_reps_sum(self._spg_reps)  # type: ignore
+        coset_reps_sum = get_compr_coset_projector_O1(self._spg_reps)  # type: ignore
         proj_rt = coset_reps_sum
 
         if len(proj_rt.data) == 0:
