+++ tests/utils/test_utils_O4.py
@@ -1,12 +1,9 @@
 """Tests of functions in utils_O4."""
 
 import numpy as np
-import pytest
 
-from symfc.spg_reps import SpgRepsO4
 from symfc.utils.utils_O4 import (
     get_atomic_lat_trans_decompr_indices_O4,
-    get_compr_coset_projector_O4,
     get_lat_trans_compr_matrix_O4,
     get_lat_trans_decompr_indices_O4,
 )
@@ -30,13 +27,3 @@ def test_lat_trans(cell_spg_reps_bcc):
     row, col = c_trans.nonzero()
     np.testing.assert_array_equal(decompr_idx, col)
     np.testing.assert_allclose(c_trans.data, [0.7071067811865475] * len(decompr_idx))
-
-
-def test_coset_projector_O4(cell_spg_reps_bcc):
-    """Test get_compr_coset_projector_O4."""
-    supercell, trans_perms, _ = cell_spg_reps_bcc
-    spg_reps = SpgRepsO4(supercell)
-    atomic_decompr_idx = get_atomic_lat_trans_decompr_indices_O4(trans_perms)
-    coset = get_compr_coset_projector_O4(spg_reps, atomic_decompr_idx)
-    assert coset.trace() == pytest.approx(32.0)
-    assert np.sum(coset.data) == pytest.approx(168.0)
